Start to remove ic2

+++ src/main/java/gregtech/api/interfaces/IItemContainer.java
@@ -13,15 +13,15 @@ public interface IItemContainer {
 
     boolean isStackEqual(ItemStack aStack, boolean aWildcard, boolean aIgnoreNBT);
 
-    ItemStack get(long aAmount, ItemStack... aReplacements);
+    ItemStack get(int aAmount, ItemStack... aReplacements);
 
-    ItemStack getWildcard(long aAmount, ItemStack... aReplacements);
+    ItemStack getWildcard(int aAmount, ItemStack... aReplacements);
 
-    ItemStack getUndamaged(long aAmount, ItemStack... aReplacements);
+    ItemStack getUndamaged(int aAmount, ItemStack... aReplacements);
 
-    ItemStack getAlmostBroken(long aAmount, ItemStack... aReplacements);
+    ItemStack getAlmostBroken(int aAmount, ItemStack... aReplacements);
 
-    ItemStack getWithDamage(long aAmount, long aMetaValue, ItemStack... aReplacements);
+    ItemStack getWithDamage(int aAmount, int aMetaValue, ItemStack... aReplacements);
 
     IItemContainer set(Item aItem);
 
@@ -31,9 +31,9 @@ public interface IItemContainer {
 
     IItemContainer registerWildcardAsOre(Object... aOreNames);
 
-    ItemStack getWithCharge(long aAmount, int aEnergy, ItemStack... aReplacements);
+    ItemStack getWithCharge(int aAmount, int aEnergy, ItemStack... aReplacements);
 
-    ItemStack getWithName(long aAmount, String aDisplayName, ItemStack... aReplacements);
+    ItemStack getWithName(int aAmount, String aDisplayName, ItemStack... aReplacements);
 
     boolean hasBeenSet();
 }
